remove-route-groups
