--- apps/web/app/future/(shared-page-wrapper)/(admin-layout)/settings/admin/oAuth/page.tsx
@@ -1,10 +1,0 @@
-import Page from "@pages/settings/admin/oAuth/index";
-import { _generateMetadata } from "app/_utils";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    () => "OAuth",
-    () => "Add new OAuth Clients"
-  );
-
-export default Page;
