--- apps/web/app/future/(shared-page-wrapper)/(admin-layout)/settings/admin/page.tsx
@@ -1,10 +1,0 @@
-import Page from "@pages/settings/admin/index";
-import { _generateMetadata } from "app/_utils";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    () => "Admin",
-    () => "admin_description"
-  );
-
-export default Page;
