+++ apps/web/app/future/apps/[slug]/layout.tsx
