+++ apps/web/app/future/apps/installed/[category]/layout.tsx
