+++ apps/web/app/future/apps/installed/[category]/page.tsx
