+++ apps/web/app/future/settings/admin/users/[id]/edit/page.tsx
