+++ apps/web/app/future/video/[uid]/page.tsx
